refacto: settings pages
- Added setting layout mobile widget.
- Added widget for QR code page.
- Added widget for regular form page.
- Refacto OBS setting page
